+++ src/main/java/com/lmalvarez/tools/security/Hash.java
@@ -1,0 +1,88 @@
+package com.lmalvarez.tools.security;
+
+import java.math.BigInteger;
+import java.nio.charset.StandardCharsets;
+import java.security.MessageDigest;
+import java.security.NoSuchAlgorithmException;
+
+public class Hash {
+    public static String MD5(String input) {
+        try {
+            MessageDigest md = MessageDigest.getInstance("MD5");
+            byte[] messageDigest = md.digest(input.getBytes());
+            BigInteger number = new BigInteger(1, messageDigest);
+            String hashtext = number.toString(16);
+            while (hashtext.length() < 32) {
+                hashtext = "0" + hashtext;
+            }
+            return hashtext;
+        } catch (NoSuchAlgorithmException e) {
+            throw new RuntimeException(e);
+        }
+    }
+
+    public static String SHA1(String input) {
+        try {
+            MessageDigest mDigest = MessageDigest.getInstance("SHA1");
+            byte[] result = mDigest.digest(input.getBytes());
+            StringBuilder sb = new StringBuilder();
+            for (int i = 0; i < result.length; i++) {
+                sb.append(Integer.toString((result[i] & 0xff) + 0x100, 16).substring(1));
+            }
+            return sb.toString();
+        } catch (NoSuchAlgorithmException ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+
+    public static String SHA256(String input) {
+        try {
+            MessageDigest sha256 = MessageDigest.getInstance("SHA-256");
+            sha256.update(input.getBytes("UTF-8"));
+            byte[] digest = sha256.digest();
+            StringBuffer sb = new StringBuffer();
+            for (int i = 0; i < digest.length; i++) {
+                sb.append(String.format("%02x", digest[i]));
+            }
+            return sb.toString();
+        } catch (Exception ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+
+    public static String SHA384(String message) {
+        try {
+            byte[] buffer = message.getBytes();
+            MessageDigest md = MessageDigest.getInstance("SHA-384");
+            md.update(buffer);
+            byte[] digest = md.digest();
+            String hexValue = null;
+            for (int i = 0; i < digest.length; i++) {
+                int b = digest[i] & 0xff;
+                if (Integer.toHexString(b).length() == 1) {
+                    hexValue = hexValue == null ? "0" : hexValue + "0";
+                }
+                hexValue = hexValue == null ? Integer.toHexString(b) : hexValue + Integer.toHexString(b);
+            }
+            return hexValue;
+        } catch (NoSuchAlgorithmException e) {
+            e.printStackTrace();
+        }
+        return null;
+    }
+
+    public static String SHA512(String input) {
+        try {
+            MessageDigest md = MessageDigest.getInstance("SHA-512");
+            byte[] bytes = md.digest(input.getBytes(StandardCharsets.UTF_8));
+            StringBuilder sb = new StringBuilder();
+            for (int i = 0; i < bytes.length; i++) {
+                sb.append(Integer.toString((bytes[i] & 0xff) + 0x100, 16).substring(1));
+            }
+            String generatedPassword = sb.toString();
+            return generatedPassword;
+        } catch (Exception ex) {
+            throw new RuntimeException(ex);
+        }
+    }
+}